Refactoring shared functions of ndimage into single util file.
