+++ cupyx/scipy/ndimage/_interp_kernels.py
@@ -1,7 +1,7 @@
 import cupy
 import cupy.core.internal
 
-from cupyx.scipy.ndimage import filters
+from cupyx.scipy.ndimage import _util
 
 
 def _get_coord_map(ndim):
@@ -237,7 +237,7 @@ def _generate_interp_custom(coord_func, ndim, large_int, yshape, mode, cval,
             if mode != 'constant':
                 ixvar = 'cf_{j}'.format(j=j)
                 ops.append(
-                    filters._generate_boundary_condition_ops(
+                    _util._generate_boundary_condition_ops(
                         mode, ixvar, 'xsize_{}'.format(j)))
 
             # sum over ic_j will give the raveled coordinate in the input
@@ -265,11 +265,11 @@ def _generate_interp_custom(coord_func, ndim, large_int, yshape, mode, cval,
             if mode != 'constant':
                 ixvar = 'cf_bounded_{j}'.format(j=j)
                 ops.append(
-                    filters._generate_boundary_condition_ops(
+                    _util._generate_boundary_condition_ops(
                         mode, ixvar, 'xsize_{}'.format(j)))
                 ixvar = 'cc_bounded_{j}'.format(j=j)
                 ops.append(
-                    filters._generate_boundary_condition_ops(
+                    _util._generate_boundary_condition_ops(
                         mode, ixvar, 'xsize_{}'.format(j)))
 
             ops.append("""
